Bump version to 1.0.1 and change LimitCheck fields from Int to Double
- Update versionCode from 1 to 2 and versionName from 1.0.0 to 1.0.1
- Change LimitCheck.used and LimitCheck.limit from Int to Double to match backend API response format
- Convert Double values to Int in UsageScreen when displaying API call limits and usage

+++ fyndable-admin-app/app/src/main/java/com/fyndable/admin/data/remote/Dtos.kt
@@ -210,8 +210,8 @@ data class LimitChecks(
 
 @Serializable
 data class LimitCheck(
-    val used: Int = 0,
-    val limit: Int = 0,
+    val used: Double = 0.0,
+    val limit: Double = 0.0,
     val exceeded: Boolean = false,
 )
